feat(CB2-11360): mved to processor
